Added a ControllerManager and updated the scene to use it
The new manager acts as a list of active devices. It currently only holds links to 2 controllers.
To get a controller on the list simply press a button on it (this includes the keyboard which are QEU or O)
TODO Change the manager so game objects can subscribe to a controller and thus have a dynamic amount of controllers.